// src/app/lib/riot/types.ts
export type TeamId = 100 | 200;

export interface Summoner {
  platformId: string;
  accountId: string;
  profileIconId?: number;
  revisionDate: number;
  name: string;
  id: string;
  puuid: string;
  summonerLevel: number;
}

export type Position = {
  x: number;
  y: number;
};
// ...
export type ItemPurchasedEvent = {
  itemId: number;
  participantId: number;
  timestamp: number;
  type: 'ITEM_PURCHASED';
};

export type LevelUpEvent = {
  level: number;
  participantId: number;
  timestamp: number;
  type: 'LEVEL_UP';
};

export type SkillLevelUpEvent = {
  levelUpType: string;
  participantId: number;
  skillSlot: number;
  timestamp: number;
  type: 'SKILL_LEVEL_UP';
};

export type WardPlacedEvent = {
  creatorId: number;
  timestamp: number;
  type: 'WARD_PLACED';
  wardType: 'YELLOW_TRINKET' | 'SIGHT_WARD' | 'CONTROL_WARD' | 'BLUE_TRINKET';
};

export type WardKillEvent = {
  killerId: number;
  timestamp: number;
  type: 'WARD_KILL';
  wardType: 'YELLOW_TRINKET' | 'SIGHT_WARD' | 'CONTROL_WARD' | 'BLUE_TRINKET';
};

export type ItemDestroyedEvent = {
  itemId: number;
  participantId: number;
  timestamp: number;
  type: 'ITEM_DESTROYED';
};

export type BuildingKillEvent = {
  assistingParticipantIds?: number[];
  buildingType: 'TOWER_BUILDING' | 'INHIBITOR_BUILDING';
  killerId: number;
  laneType: 'TOP_LANE' | 'MID_LANE' | 'BOT_LANE';
  position: Position;
  teamId: TeamId;
  timestamp: number;
  towerType?: 'OUTER_TURRET' | 'INNER_TURRET' | 'BASE_TURRET' | 'NEXUS_TURRET';
  type: 'BUILDING_KILL';
};

export type TurretPlateDestroyedEvent = {
  killerId: number;
  laneType: 'TOP_LANE' | 'MID_LANE' | 'BOT_LANE';
  position: Position;
  teamId: TeamId;
  timestamp: number;
  type: 'TURRET_PLATE_DESTROYED';
};

export type ChampionKillEvent = {
  assistingParticipantIds?: number[];
  bounty: number;
  killStreakLength: number;
  killerId: number;
  position: Position;
  timestamp: number;
  type: 'CHAMPION_KILL';
  victimDamageDealt: {
    basic: boolean;
    magicDamage: number;
    name: string;
    participantId: number;
    physicalDamage: number;
    spellName: string;
    spellSlot: number;
    trueDamage: number;
    type: string;
  }[];
  victimDamageReceived: {
    basic: boolean;
    magicDamage: number;
    name: string;
    participantId: number;
    physicalDamage: number;
    spellName: string;
    spellSlot: number;
    trueDamage: number;
    type: string;
  }[];
  victimId: number;
};

export type PauseEndEvent = {
  realTimestamp: number;
  timestamp: number;
  type: 'PAUSE_END';
};

export type ChampionSpecialKillEvent = {
  killType: 'KILL_FIRST_BLOOD' | 'KILL_MULTI';
  killerId: number;
  multiKillLength?: number;
  position: Position;
  timestamp: number;
  type: 'CHAMPION_SPECIAL_KILL';
};

export type GameEndEvent = {
  gameId: number;
  realTimestamp: number;
  timestamp: number;
  type: 'GAME_END';
  winningTeam: number;
};

export type EliteMonsterKill = {
  assistingParticipantIds?: number[];
  killerId: number;
  killerTeamId: TeamId;
  monsterSubType?:
    | 'FIRE_DRAGON'
    | 'EARTH_DRAGON'
    | 'AIR_DRAGON'
    | 'WATER_DRAGON'
    | 'ELDER_DRAGON';
  monsterType: 'RIFTHERALD' | 'DRAGON' | 'BARON_NASHOR';
  position: Position;
  timestamp: number;
  type: 'ELITE_MONSTER_KILL';
};

export type MatchEvent =
  | PauseEndEvent
  | ItemPurchasedEvent
  | LevelUpEvent
  | SkillLevelUpEvent
  | WardPlacedEvent
  | WardKillEvent
  | ItemDestroyedEvent
  | BuildingKillEvent
  | TurretPlateDestroyedEvent
  | ChampionKillEvent
  | ChampionSpecialKillEvent
  | GameEndEvent
  | EliteMonsterKill;